Moved guestnetwork to communication folder
